refactor(WinnerCalculator.java): 우승자 로직 분리
Application 클래스가 이름 생성과 우승자 계산이라는 두 가지 기능을 가지고 있어 각각의 역할을 명확히 하기 위해 WinnerCalculator로 분리했습니다
